src/requirements.txt
rich==13.7.0
python-docx>=0.8.11
PyPDF2==3.0.1
spacy>=3.7.2
jinja2==3.1.3
python-dateutil==2.8.2
pyinstaller==6.3.0
docx2pdf>=0.1.8
psutil>=5.9.0
Pillow>=10.0.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl 